+++ src/g1.rs
@@ -257,9 +257,57 @@ impl G1Affine {
         }
     }
 
+    /// Raw bytes representation
+    ///
+    /// The intended usage of this function is for trusted sets of data where performance is
+    /// critical. This way, the `infinity` internal attribute will not be stored and `x` and
+    /// `y` will be stored without any check.
+    ///
+    /// For secure serialization, check `to_compressed` and `to_uncompressed`
+    pub unsafe fn to_bytes_unchecked(&self) -> [u8; 96] {
+        let mut bytes = [0u8; 96];
+        let chunks = bytes.chunks_mut(8);
+
+        self.x
+            .internal_repr()
+            .iter()
+            .chain(self.y.internal_repr().iter())
+            .zip(chunks)
+            .for_each(|(n, c)| c.copy_from_slice(&n.to_le_bytes()));
+
+        bytes
+    }
+
+    /// Create a `G1Affine` from a set of bytes created by `to_bytes_unchecked`.
+    ///
+    /// No check is performed and no constant time is granted. The `infinity` attribute is also
+    /// lost. The expected usage of this function is for trusted bytes where performance is
+    /// critical.
+    ///
+    /// For secure serialization, check `to_bytes`
+    pub unsafe fn from_slice_unchecked(bytes: &[u8]) -> Self {
+        let mut x = [0u64; 6];
+        let mut y = [0u64; 6];
+        let mut z = [0u8; 8];
+
+        bytes
+            .chunks_exact(8)
+            .zip(x.iter_mut().chain(y.iter_mut()))
+            .for_each(|(c, n)| {
+                z.copy_from_slice(c);
+                *n = u64::from_le_bytes(z);
+            });
+
+        let x = Fp::from_raw_unchecked(x);
+        let y = Fp::from_raw_unchecked(y);
+        let infinity = 0u8.into();
+
+        Self { x, y, infinity }
+    }
+
     /// Serializes this element into compressed form. See [`notes::serialization`](crate::notes::serialization)
     /// for details about how group elements are serialized.
-    pub fn to_compressed(&self) -> [u8; 48] {
+    pub fn to_bytes(&self) -> [u8; 48] {
         // Strictly speaking, self.x is zero already when self.infinity is true, but
         // to guard against implementation mistakes we do not assume this.
         let mut res = Fp::conditional_select(&self.x, &Fp::zero(), self.infinity).to_bytes();
@@ -282,101 +330,12 @@ impl G1Affine {
         res
     }
 
-    /// Serializes this element into uncompressed form. See [`notes::serialization`](crate::notes::serialization)
-    /// for details about how group elements are serialized.
-    pub fn to_uncompressed(&self) -> [u8; 96] {
-        let mut res = [0; 96];
-
-        res[0..48].copy_from_slice(
-            &Fp::conditional_select(&self.x, &Fp::zero(), self.infinity).to_bytes()[..],
-        );
-        res[48..96].copy_from_slice(
-            &Fp::conditional_select(&self.y, &Fp::zero(), self.infinity).to_bytes()[..],
-        );
-
-        // Is this point at infinity? If so, set the second-most significant bit.
-        res[0] |= u8::conditional_select(&0u8, &(1u8 << 6), self.infinity);
-
-        res
-    }
-
-    /// Attempts to deserialize an uncompressed element. See [`notes::serialization`](crate::notes::serialization)
-    /// for details about how group elements are serialized.
-    pub fn from_uncompressed(bytes: &[u8; 96]) -> CtOption<Self> {
-        Self::from_uncompressed_unchecked(bytes)
-            .and_then(|p| CtOption::new(p, p.is_on_curve() & p.is_torsion_free()))
-    }
-
-    /// Attempts to deserialize an uncompressed element, not checking if the
-    /// element is on the curve and not checking if it is in the correct subgroup.
-    /// **This is dangerous to call unless you trust the bytes you are reading; otherwise,
-    /// API invariants may be broken.** Please consider using `from_uncompressed()` instead.
-    pub fn from_uncompressed_unchecked(bytes: &[u8; 96]) -> CtOption<Self> {
-        // Obtain the three flags from the start of the byte sequence
-        let compression_flag_set = Choice::from((bytes[0] >> 7) & 1);
-        let infinity_flag_set = Choice::from((bytes[0] >> 6) & 1);
-        let sort_flag_set = Choice::from((bytes[0] >> 5) & 1);
-
-        // Attempt to obtain the x-coordinate
-        let x = {
-            let mut tmp = [0; 48];
-            tmp.copy_from_slice(&bytes[0..48]);
-
-            // Mask away the flag bits
-            tmp[0] &= 0b0001_1111;
-
-            Fp::from_bytes(&tmp)
-        };
-
-        // Attempt to obtain the y-coordinate
-        let y = {
-            let mut tmp = [0; 48];
-            tmp.copy_from_slice(&bytes[48..96]);
-
-            Fp::from_bytes(&tmp)
-        };
-
-        x.and_then(|x| {
-            y.and_then(|y| {
-                // Create a point representing this value
-                let p = G1Affine::conditional_select(
-                    &G1Affine {
-                        x,
-                        y,
-                        infinity: infinity_flag_set,
-                    },
-                    &G1Affine::identity(),
-                    infinity_flag_set,
-                );
-
-                CtOption::new(
-                    p,
-                    // If the infinity flag is set, the x and y coordinates should have been zero.
-                    ((!infinity_flag_set) | (infinity_flag_set & x.is_zero() & y.is_zero())) &
-                    // The compression flag should not have been set, as this is an uncompressed element
-                    (!compression_flag_set) &
-                    // The sort flag should not have been set, as this is an uncompressed element
-                    (!sort_flag_set),
-                )
-            })
-        })
-    }
-
     /// Attempts to deserialize a compressed element. See [`notes::serialization`](crate::notes::serialization)
     /// for details about how group elements are serialized.
-    pub fn from_compressed(bytes: &[u8; 48]) -> CtOption<Self> {
+    pub fn from_bytes(bytes: &[u8; 48]) -> CtOption<Self> {
         // We already know the point is on the curve because this is established
         // by the y-coordinate recovery procedure in from_compressed_unchecked().
 
-        Self::from_compressed_unchecked(bytes).and_then(|p| CtOption::new(p, p.is_torsion_free()))
-    }
-
-    /// Attempts to deserialize an uncompressed element, not checking if the
-    /// element is in the correct subgroup.
-    /// **This is dangerous to call unless you trust the bytes you are reading; otherwise,
-    /// API invariants may be broken.** Please consider using `from_compressed()` instead.
-    pub fn from_compressed_unchecked(bytes: &[u8; 48]) -> CtOption<Self> {
-        // Obtain the three flags from the start of the byte sequence
         let compression_flag_set = Choice::from((bytes[0] >> 7) & 1);
         let infinity_flag_set = Choice::from((bytes[0] >> 6) & 1);
         let sort_flag_set = Choice::from((bytes[0] >> 5) & 1);
@@ -428,6 +387,7 @@ impl G1Affine {
                 })
             })
         })
+        .and_then(|p| CtOption::new(p, p.is_torsion_free()))
     }
 
     /// Returns true if this element is the identity (the point at infinity).
@@ -1516,4 +1476,16 @@ fn g1_affine_serde_roundtrip() {
     let deser: G1Affine = bincode::deserialize(&ser).unwrap();
 
     assert_eq!(gen, deser);
+}
+
+#[test]
+fn g1_affine_bytes_unchecked() {
+    let gen = G1Affine::generator();
+    let gen_p = unsafe {
+        let bytes = gen.to_bytes_unchecked();
+
+        G1Affine::from_slice_unchecked(&bytes)
+    };
+
+    assert_eq!(gen, gen_p);
 }