+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 12-01-21
+### Changed
+- G1Affine method rename to/from bytes
+### Added
+- Included `G1Affine::to_bytes_unchecked`
+
 ## [0.4.0] - 24-12-20
 ### Changed
 - no-std compatibility for pairings feature